Safe algorithm tested & working
Making this commit to save the current state of the program because it's in a good spot for testing the safe state function. It seems to be working with several different test inputs but I'd like to be able to come back if it starts having issues once I'm rolling with 18 random user processes. Next: User behavior and message queue