Added most functionality
It's working but math is messing up somewhere and/or resources are being added when they shouldn't. I have a lot of code to go through to figure out where it's failing. Currently have lots of output to keep track of what's happening.
Still need to add logging, I'll worry about that last.